refactor(iocuak): move mocks to containerModuleMetadata folder
